Added The Ideal Team Player
I just finished reading most of _The Ideal Team Player_ and feel it is a worthy successor to _The Five Dysfunctions of a Team_.  Whereas the older book focused on the team as a whole, this new one focuses on the virtues of the individual team members and what makes someone a great fit for a team.  It builds a bit on a couple of the things talked about in Jim Collins' _Good to Great_ as well.